requirements.txt
flask
torch
torchvision
transformers
Pillow
flask-cors
python-dotenv
requests